# services/backend/blue-magma-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for blue magma auth and data",
        "title": "Bluemagma API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/api/v1/org/{org_id}/codebase": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all codebases for an organization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "codebase"
                ],
                "summary": "Get all codebases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.CodebaseResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new codebase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "codebase"
                ],
                "summary": "Create a new codebase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service data",
                        "name": "codebase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CodebaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CodebaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/api/v1/org/{org_id}/codebase/{service_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a codebase by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "codebase"
                ],
                "summary": "Get a codebase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Codebase ID",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CodebaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a codebase by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "codebase"
                ],
                "summary": "Update a codebase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Codebase ID",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Codebase data",
                        "name": "codebase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CodebaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CodebaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a codebase by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "codebase"
                ],
                "summary": "Delete a codebase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Codebase ID",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
        "handlers.CodebaseRequest": {
            "type": "object",
            "properties": {
                "codebase_description": {
                    "type": "string"
                },
                "codebase_name": {
                    "type": "string"
                },
                "codebase_repo_url": {
                    "type": "string"
                },
                "codebase_type": {
                    "description": "Optional field for codebase type",
                    "type": "string"
                },
                "github_installation_id": {
                    "type": "integer"
                },
                "github_repo_id": {
                    "type": "integer"
                },
                "source_type": {
                    "description": "GitHub integration fields",
                    "type": "string"
                }
            }
        },
        "handlers.CodebaseResponse": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "codebase_description": {
                    "type": "string"
                },
                "codebase_name": {
                    "type": "string"
                },
                "codebase_repo_url": {
                    "type": "string"
                },
                "codebase_type": {
                    "description": "Optional field for codebase type",
                    "type": "string"
                },
                "object_id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "versions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.serviceVersionResponse"
                    }
                }
            }
        },
...
        "handlers.serviceVersionResponse": {
            "type": "object",
            "properties": {
                "branch_name": {
                    "type": "string"
                },
                "commit_hash": {
                    "type": "string"
                },
                "object_id": {
                    "type": "string"
                },
                "summary": {
                    "description": "Optional field for summary",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}